Do not count empty days for average per day
This is useful if you just downloaded the app and, e.g., have just 3
Days in the 30 Day view which contain anything. This way, the sleep is
for example set to 8hrs a day and not ~50Minutes.

The coreceAtLeast should not be needed as the division with totalSeconds
only happens if there is one interval, but to be sure (e.g, a 0 second
interval or something) i added it.